--- pyRevit/__RPS__userInit.py
@@ -1,45 +1,0 @@
-'''
-Copyright (c) 2014-2016 Ehsan Iran-Nejad
-Python scripts for Autodesk Revit
-
-This file is part of pyRevit repository at https://github.com/eirannejad/pyRevit
-
-pyRevit is a free set of scripts for Autodesk Revit: you can redistribute it and/or modify
-it under the terms of the GNU General Public License version 3, as published by
-the Free Software Foundation.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-See this link for a copy of the GNU General Public License protecting this package.
-https://github.com/eirannejad/pyRevit/blob/master/LICENSE
-'''
-
-# these commands get executed in the current scope
-# of each new shell (but not for canned commands)
-from Autodesk.Revit.DB import *
-from Autodesk.Revit.DB.Architecture import *
-from Autodesk.Revit.DB.Analysis import *
-from Autodesk.Revit.UI import TaskDialog
-from Autodesk.Revit.UI import UIApplication
-
-uidoc = __revit__.ActiveUIDocument
-doc = __revit__.ActiveUIDocument.Document
-selection = [ doc.GetElement( elId ) for elId in __revit__.ActiveUIDocument.Selection.GetElementIds() ]
-
-def alert(msg):
-	TaskDialog.Show('RevitPythonShell', msg)
-
-def quit():
-	__window__.Close()
-
-def free():
-	__window__.Hide()
-	__window__.Show()
-
-exit = quit
-
-if len(selection) > 0:
-	el = selection[0]
